extract outputformat
